Additional Functions

Useful or Helper Functions

Here are functions that have a usecase in other areas that fall not strictlly into any offical categories for trend estimation.

TrendDecomposition.greatestConvexMinorant — Function

greatestConvexMinorant(y :: Vector)

Computes greatest convex minorant of series y using the pool-adjecent-violators algorithm

Returns the coordinates of the knots as a tuple (x, y)

TrendDecomposition.leastConcaveMajorant — Method

leastConcaveMajorant(y :: Vector)

Computes least concave majorant (lcm) of series y using the pool-adjecent-violators algorithm

Returns the coordinates of the knots as a tuple (x, y)

Experimental

Here are functions that are only experimental in use but could be useful for some specific purposes.

TrendDecomposition.trendL1Filter — Function

trendL1Filter(y :: Vector, λ :: Real; m = 2, max_iter=20, method = :ADMM)

Placeholder for trendL1Filter extension, when using TrendDecomposition together with other Julia packages like Convex.jl and SCS.jl.

This function provides the generic use of serveral optimization methods to compute a numerical solution. Following methods are implmented: :ADMM -> alternating direction method of multipliers :ConvexSCS -> SCS solver with Convex.jl. Prerequisite! Import necessary modules with: using Convex, SCS

The function returns the estimated trend component